requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==3.1.1         # via google-auth
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
google-api-core[grpc]==1.14.3  # via google-cloud-core, google-cloud-datastore
google-auth==1.6.3        # via google-api-core
google-cloud-core==1.0.3  # via google-cloud-datastore
google-cloud-datastore==1.10.0  # via google-cloud-ndb
google-cloud-ndb==0.1.0
googleapis-common-protos==1.6.0  # via google-api-core
grpcio==1.24.3            # via google-api-core
idna==2.8                 # via requests
protobuf==3.10.0          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.7     # via google-auth
pyasn1==0.4.7             # via pyasn1-modules, rsa
pytz==2019.3              # via google-api-core
redis==3.3.11             # via google-cloud-ndb
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-auth, grpcio, protobuf
urllib3==1.25.6           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.4.0        # via google-api-core, protobuf
